Type refactor
