+++ package.json
@@ -19,8 +19,8 @@
   "main": "./dist/bundles/meilisearch.umd.js",
   "module": "./dist/bundles/meilisearch.esm.js",
   "browser": "./dist/bundles/meilisearch.umd.js",
-  "typings": "./dist/types/types.d.ts",
-  "types": "./dist/types/types.d.ts",
+  "typings": "./dist/types/index.d.ts",
+  "types": "./dist/types/index.d.ts",
   "sideEffects": false,
   "repository": {
     "type": "git",
