+++ scripts/test.js
@@ -1,0 +1,10 @@
+const a = [
+  [
+    [
+      1,2,3,4
+    ]
+  ]
+]
+
+
+console.log(a)
