+++ src/errors/index.ts
@@ -1,0 +1,5 @@
+export * from './http-error-handler'
+export * from './meilisearch-api-error'
+export * from './meilisearch-communication-error'
+export * from './meilisearch-error'
+export * from './meilisearch-timeout-error'
