+++ tsconfig.json
@@ -7,8 +7,11 @@
       "declaration": true,
       "declarationMap": true,
       "declarationDir": "./dist/types",
+
       "allowSyntheticDefaultImports": true,
       "lib": ["ES2019", "ESNext", "dom"],
+
+      "moduleResolution": "node",
       "noImplicitAny": true,
       "strictNullChecks": true,
       "strictFunctionTypes": true,
